+++ tests/testthat/test_ETA.R
@@ -20,12 +20,15 @@ d <- dist(xy)
 
 x <- defineTrajectories(d, sites, surveys, times)
 
+
 test_that("Trajectories can be analyzed",{
   expect_s3_class(trajectoryLengths(x), "data.frame")
   expect_s3_class(trajectoryLengths(x, relativeToInitial = TRUE), "data.frame")
   expect_s3_class(trajectoryLengths(x, all = TRUE), "data.frame")
   expect_s3_class(trajectoryLengths(x, relativeToInitial = TRUE, all = TRUE), "data.frame")
+  expect_equal(trajectoryLengths(x), trajectoryLengths2D(xy, sites, surveys))
   expect_s3_class(trajectorySpeeds(x), "data.frame")
+  expect_equal(trajectorySpeeds(x), trajectorySpeeds2D(xy, sites, surveys, times))
   expect_s3_class(trajectoryAngles(x), "data.frame")
   expect_s3_class(trajectoryAngles(x, relativeToInitial = TRUE), "data.frame")
   expect_s3_class(trajectoryAngles(x, all = TRUE), "data.frame")
@@ -35,4 +38,32 @@ test_that("Trajectories can be analyzed",{
   expect_type(segmentDistances(x), "list")
   expect_s3_class(trajectoryDistances(x), "dist")
   expect_type(trajectoryConvergence(x), "list")
+})
+
+# Shuffle surveys and check if the results are the same
+xy2 <- xy
+surveys2 <- surveys
+times2  <- times
+sites2 <- sites
+temp = xy2[5,]
+xy2[5,] = xy2[6,]
+xy2[6,] = temp
+surveys2[5] = 2
+surveys2[6] = 1
+temp = times2[5]
+times2[5] = times2[6]
+times2[6] = temp
+x2 <- defineTrajectories(dist(xy2), sites2, surveys2, times2)
+
+test_that("Trajectory analysis gives the same result after shuffling surveys",{
+  expect_equal(trajectoryLengths(x), trajectoryLengths(x2))
+  expect_equal(trajectoryLengths2D(xy, sites, surveys), trajectoryLengths2D(xy2, sites2, surveys2))
+  expect_equal(trajectorySpeeds(x), trajectorySpeeds(x2))
+  expect_equal(trajectorySpeeds2D(xy, sites, surveys, times), trajectorySpeeds2D(xy2, sites2, surveys2, times2))
+  expect_equal(trajectoryAngles(x), trajectoryAngles(x2))
+  expect_equal(trajectoryDirectionality(x), trajectoryDirectionality(x2))
+  expect_equal(trajectoryVariability(x), trajectoryVariability(x2))
+  expect_equal(segmentDistances(x), segmentDistances(x2))
+  expect_equal(trajectoryDistances(x), trajectoryDistances(x2))
+  expect_equal(trajectoryConvergence(x), trajectoryConvergence(x2))
 })